+++ src-test/flatcc-test.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 set -e
 
-flatccCmd=./../../../../../../../flatcc/bin/flatcc
+flatccCmd=flatcc
 ${flatccCmd} --version
 ${flatccCmd} --common --builder c_test.fbs