+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.3.0
+-----
+* ARM support (v6 and v7)
+* New typedefs : obx_id, obx_schema_id, obx_uid, obx_err
+* Added obx_cursor_get_all as an alternative to iterating over all objects
+
 0.2.0
 -----
 * Added queries (obx_qb_* and obx_query_*)